Mark DispatchTimeInterval Equatable
We can mark `DispatchTimeInterval` as equatable because it already
implements the equality check operator `==`.

Doing this will allow DispatchTimeInterval to be used with a wide range
of library functions which are defined to operate on Equatable objects.
For example this is useful in tests when you might which to use `XCTAssertEqual`.