build: slightly improve the build for libdispatch
Avoid polluting the build directory a small amount given that we can use
`-fmodule-map-file=` for the C/C++ build of libdispatch.  Unfortunately,
for the Swift build, we need to have the file copied over due to the
umbrella header resolution.

Hopefully this reduces some of the race conditions that we have seen in
the build.

Thanks to @dgregor for reminding me of the flag!