dispatch: install the correct modulemap when building static
There is a subtle difference between libdispatch built dynamically and
statically: DispatchStubs.  We erroneously emit ObjC runtime calls into
the build and the stubs provides a shim to provide a definition on
non-ObjC runtime enabled targets.  When built dynamically, this is
internalised and distributed as part of dispatch.dll/libdispatch.so,
however when built statically, the consumer is responsible for linking
against DispatchStubs.  The modulemap reflects this and we need to
ensure that we correctly provide that modulemap.

Thanks to @MaxDesiatov for the help with debugging and testing a fix for
this!

Fixes: rdar://23335318